refactor(test): move ITimerService fake to Support
