+++ tests/Moongate.Tests/Server/Game/GameLoopContextTests.cs
@@ -1,4 +1,5 @@
 using Moongate.Server.Services.Game;
+using Moongate.Tests.Support;
 using SquidStd.Services.Core.Services;
 
 namespace Moongate.Tests.Server.Game;
@@ -29,34 +30,4 @@ public class GameLoopContextTests
 
         await Assert.ThrowsAsync<TimeoutException>(async () => await task);
     }
-}
-
-/// <summary>
-/// Stands in for the real timer wheel: <see cref="GameLoopContext" /> only needs an
-/// <see cref="SquidStd.Core.Interfaces.Timing.ITimerService" /> to satisfy its constructor for these
-/// tests, which exercise <see cref="GameLoopContext.InvokeAsync{T}" /> and never touch timers.
-/// </summary>
-internal sealed class FakeTimerServiceForContext : SquidStd.Core.Interfaces.Timing.ITimerService
-{
-    public string RegisterTimer(
-        string name,
-        System.TimeSpan interval,
-        System.Action callback,
-        System.TimeSpan? delay = null,
-        bool repeat = false
-    )
-        => name;
-
-    public void UnregisterAllTimers()
-    {
-    }
-
-    public bool UnregisterTimer(string timerId)
-        => true;
-
-    public int UnregisterTimersByName(string name)
-        => 0;
-
-    public int UpdateTicksDelta(long timestampMilliseconds)
-        => 0;
 }
